+++ route/help.go
@@ -51,6 +51,6 @@ func (b *Kit) GetCommandInfo() []*CommandInfo {
 }
 
 // GetCommandInfo returns information about all commands in the kit (intended for use in help commands)
-func (m *MessageContext) GetCommandInfo() []*CommandInfo {
-	return m.Kit.GetCommandInfo()
+func (m *CommonContext) GetCommandInfo() []*CommandInfo {
+	return m.kit.GetCommandInfo()
 }