+++ route/run.go
@@ -202,7 +202,7 @@ func (b *Kit) onMessageCreate(session *discordgo.Session, message *discordgo.Mes
 	ctx := &MessageContext{
 		CommonContext: &CommonContext{
 			Session: session,
-			Kit:     b,
+			kit:     b,
 		},
 		Message:   message,
 		Raw:       trimmedContent,
@@ -229,7 +229,7 @@ func (b *Kit) onReactionAdd(session *discordgo.Session, reaction *discordgo.Mess
 	mCtx := ReactionContext{
 		CommonContext: &CommonContext{
 			Session: session,
-			Kit:     b,
+			kit:     b,
 		},
 		Reaction: reaction.MessageReaction,
 		Event:    ReactionAdd,
@@ -273,7 +273,7 @@ func (b *Kit) onReactionRemove(session *discordgo.Session, reaction *discordgo.M
 	mCtx := ReactionContext{
 		CommonContext: &CommonContext{
 			Session: session,
-			Kit:     b,
+			kit:     b,
 		},
 		Reaction: reaction.MessageReaction,
 		Event:    ReactionRemove,